transfer query to db in background
