+++ app/src/main/java/org/stepic/droid/concurrency/LoadingCoursesTask.java
@@ -1,18 +1,17 @@
 package org.stepic.droid.concurrency;
 
-import android.content.Context;
-
 import org.stepic.droid.base.MainApplication;
 import org.stepic.droid.core.IShell;
-import org.stepic.droid.exceptions.NullCourseListException;
 import org.stepic.droid.model.Course;
+import org.stepic.droid.store.operations.DbOperationsCourses;
 import org.stepic.droid.web.IApi;
 
+import java.sql.SQLException;
 import java.util.List;
 
 import javax.inject.Inject;
 
-public class LoadingCoursesTask extends StepicTask <Void, Void, List<Course>> {
+public class LoadingCoursesTask extends StepicTask<Void, Void, List<Course>> {
 
     @Inject
     IShell mShell;
@@ -32,7 +31,6 @@ public class LoadingCoursesTask extends StepicTask <Void, Void, List<Course>> {
 
     @Override
     protected List<Course> doInBackgroundBody(Void... params) throws Exception {
-        Thread.sleep(3000); //todo: delete fake latency for debug
         IApi api = mShell.getApi();
         List<Course> courseList = null;
         switch (mCourseType) {
@@ -44,7 +42,59 @@ public class LoadingCoursesTask extends StepicTask <Void, Void, List<Course>> {
                 break;
         }
 
-        if (courseList == null) throw new NullCourseListException();
+        if (courseList != null)
+        {
+            List<Course> cachedCourses =getCachedCourses();
+
+            DbOperationsCourses dbOperationCourses = mShell.getDbOperationsCourses(getDbType(mCourseType));
+            for (Course courseItem : cachedCourses) {
+                if (!courseList.contains(courseItem)) {
+                    dbOperationCourses.deleteCourse(courseItem);//remove outdated courses from cache
+                    courseList.remove(courseItem);
+                }
+            }
+
+            for (Course newCourse : courseList) {
+                if (!dbOperationCourses.isCourseInDB(newCourse)) {
+                    dbOperationCourses.addCourse(newCourse);//add new to persistent cache
+                }
+            }
+            dbOperationCourses.close();
+            //all courses are cached now
+        }
+
+        courseList = getCachedCourses(); //get from cache;
         return courseList;
+
+    }
+    private DbOperationsCourses.Table getDbType(LoadingCoursesTask.CourseType type) {
+        DbOperationsCourses.Table dbType = null;
+        switch (type) {
+            case enrolled:
+                dbType = DbOperationsCourses.Table.enrolled;
+                break;
+            case featured:
+                dbType = DbOperationsCourses.Table.featured;
+        }
+        return dbType;
+    }
+
+    private List<Course> getCachedCourses() {
+        DbOperationsCourses dbOperationCourses = mShell.getDbOperationsCourses(getDbType(mCourseType));
+
+        try {
+            dbOperationCourses.open();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        List<Course> cachedCourses = dbOperationCourses.getAllCourses();
+        dbOperationCourses.close();
+
+        return cachedCourses;
+    }
+
+    @Override
+    protected void onPostExecute(AsyncResultWrapper<List<Course>> listAsyncResultWrapper) {
+        super.onPostExecute(listAsyncResultWrapper);
     }
 }
